move task only if different column

+++ lib/db.ts
@@ -92,16 +92,18 @@ export async function moveTask(
   taskId: string
 ) {
   try {
-    const oldTaskRef = doc(firestore, `columns/${oldColumnId}/tasks`, taskId);
-    const newTaskRef = doc(firestore, `columns/${newColumnId}/tasks`, taskId);
-    const taskSnapshot = await getDoc(oldTaskRef);
-    if (taskSnapshot.exists()) {
-      await setDoc(newTaskRef, {
-        title: taskSnapshot.data().title,
-        description: taskSnapshot.data().description,
-      });
-      await deleteDoc(oldTaskRef);
-    } else console.error("Task does not exist");
+    if (oldColumnId !== newColumnId) {
+      const oldTaskRef = doc(firestore, `columns/${oldColumnId}/tasks`, taskId);
+      const newTaskRef = doc(firestore, `columns/${newColumnId}/tasks`, taskId);
+      const taskSnapshot = await getDoc(oldTaskRef);
+      if (taskSnapshot.exists()) {
+        await deleteDoc(oldTaskRef);
+        await setDoc(newTaskRef, {
+          title: taskSnapshot.data().title,
+          description: taskSnapshot.data().description,
+        });
+      } else console.error("Task does not exist");
+    }
   } catch (error) {
     return { error };
   }
